chore(info.json): remove config/info.json file
The config/info.json file was deleted as it is no longer needed in the project. This file contained layout information for a specific component, but it is no longer used or relevant to the current state of the project. Therefore, it was decided to remove it to keep the project clean and avoid unnecessary clutter.